requirements.txt
aiogram
aioredis<2.0
redis
psycopg2
peewee
pydantic
uvicorn
fastapi
fastapi-security-telegram-webhook
fastapi-utils
python-multipart
python-jose[cryptography]
passlib[bcrypt]
beautifulsoup4
requests
google-api-python-client
validators
lxml
